+++ locales/fr.json
@@ -1427,5 +1427,44 @@
   "Import from JSON": "Importer depuis JSON",
   "Up to 20 at a time": "Jusqu'à 20 à la fois",
   "Browse community": "Explorer la communauté",
-  "Recipes shared by others": "Recettes partagées par d'autres"
+  "Recipes shared by others": "Recettes partagées par d'autres",
+
+  "Library": "Bibliothèque",
+  "Your recipes": "Vos recettes",
+  "All": "Tout",
+  "Sort": "Trier",
+  "Grid": "Grille",
+  "List": "Liste",
+  "In plan": "Au menu",
+  "Mine": "À moi",
+  "Share All": "Tout partager",
+  "Make All Private": "Tout rendre privé",
+  "Empty": "Vide",
+  "New recipe": "Nouvelle recette",
+  "Method": "Méthode",
+  "Time": "Temps",
+  "Prep": "Prépa",
+  "Serves": "Portions",
+  "For": "Pour",
+  "Adjust": "Ajuster",
+  "Advance prep": "Préparation à l'avance",
+  "Imported": "Importée",
+  "Make Private": "Rendre privée",
+  "Make Public": "Rendre publique",
+  "steps": "étapes",
+  "Create a recipe": "Créer une recette",
+  "Edit recipe": "Modifier la recette",
+  "New": "Nouvelle",
+  "Recipe name": "Nom de la recette",
+  "Origin": "Origine",
+  "Remove": "Supprimer",
+  "Describe this step...": "Décrivez cette étape...",
+  "Jasmine rice": "Riz jasmin",
+  "Drop a .json file": "Déposez un fichier .json",
+  "Choose file": "Choisir un fichier",
+  "Max 1MB · Up to 20 files · JSON only": "Max 1 Mo · Jusqu'à 20 fichiers · JSON uniquement",
+  "JSON schema": "Schéma JSON",
+  "Importing recipes…": "Importation des recettes…",
+  "Cross-checking ingredients and aisles": "Vérification des ingrédients et des rayons",
+  "This may take a moment": "Cela peut prendre un instant"
 }
