+++ templates/_settings.html
@@ -27,7 +27,6 @@
     <div class="hidden lg:block space-y-2">
       {% call nav_item(settings_path, "general", "General") %}{% endcall %}
       {% call nav_item(settings_path, "account", "Account") %}{% endcall %}
-      {% call nav_item(settings_path, "recipes", "Recipes") %}{% endcall %}
       {% call nav_item(settings_path, "billing", "Billing") %}{% endcall %}
 
       {% block sidebar_bottom %}{% endblock %}
@@ -38,7 +37,6 @@
       <div class="flex gap-2 min-w-max">
       {% call mobile_nav_item(settings_path, "general", "General") %}{% endcall %}
       {% call mobile_nav_item(settings_path, "account", "Account") %}{% endcall %}
-      {% call mobile_nav_item(settings_path, "recipes", "Recipes") %}{% endcall %}
       {% call mobile_nav_item(settings_path, "billing", "Billing") %}{% endcall %}
       </div>
     </div>
