+++ templates/_user.html
@@ -26,7 +26,7 @@
 
         <div class="flex items-center gap-4">
           <div class="flex items-center gap-1 pl-4">
-            <button ts-req="/settings/recipes/create" ts-req-method="post"
+            <button ts-req="/recipes/create" ts-req-method="post"
               class="flex items-center justify-center w-9 h-9 rounded-lg text-ink-2 hover:text-ink hover:bg-cream transition"
               title="{{ "Create Recipe"|t }}">
               <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
