+++ templates/onboarding-recipe.html
@@ -49,7 +49,7 @@
 
   <div class="space-y-2.5">
     {# Option 1 — Write your own (herb) #}
-    <button ts-req="/settings/recipes/create" ts-req-method="post"
+    <button ts-req="/recipes/create" ts-req-method="post"
       class="w-full bg-paper rounded-2xl border border-line-2 shadow-sm p-3.5 hover:bg-cream transition text-left">
       <div class="flex items-center gap-3">
         <div class="w-10 h-10 rounded-xl bg-meal-side-soft flex items-center justify-center shrink-0">
