+++ templates/partials/recipes-importing.html
@@ -1,27 +1,31 @@
 <div>
   {% for error in error_recipes %}
-  <div class="bg-red-50 border border-[#c3321a] rounded-xl p-4 mb-6">
-    <div class="flex items-start gap-3">
-      <svg class="w-5 h-5 text-[#c3321a] shrink-0 mt-0.5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-      </svg>
-      <div class="flex-1">
-        <div class="font-semibold text-[#c3321a]">{{ error.name }}</div>
-        <div class="text-sm text-[#c3321a] mt-1">{{ error.error }}</div>
-      </div>
+  <div data-import-error class="bg-red-50 border border-red-200 rounded-2xl p-4 mb-3 flex items-start gap-3">
+    <svg class="w-5 h-5 text-red-600 shrink-0 mt-0.5" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+      <path stroke-linecap="round" stroke-linejoin="round" d="M12 9v2m0 4h.01M10.29 3.86L1.82 18a2 2 0 001.71 3h16.94a2 2 0 001.71-3L13.71 3.86a2 2 0 00-3.42 0z"/>
+    </svg>
+    <div class="flex-1 min-w-0">
+      <div class="font-semibold text-sm text-red-900">{{ error.name }}</div>
+      <div class="text-xs text-red-700 mt-1 leading-relaxed">{{ error.error }}</div>
     </div>
   </div>
   {% endfor %}
   {% if let Some(id) = id %}
-  <div id="recipes-importing-alert" class="bg-yellow-50 border border-yellow-200 rounded-xl p-4 mb-6">
-    <div class="flex items-start gap-3">
-      <div class="shrink-0 mt-0.5">
-        <div class="animate-spin rounded-full h-5 w-5 border-2 border-yellow-200 border-t-yellow-600"></div>
-      </div>
-      <div class="flex-1">
-        <div class="font-semibold text-yellow-800">{{ "Importing..."|t }}</div>
-        <div class="text-sm text-yellow-700 mt-1">{{ "Creating validated recipes."|t }}</div>
-      </div>
+  <div id="recipes-importing-alert"
+    class="bg-paper border border-line-2 rounded-2xl p-5 md:p-6 flex flex-col items-center text-center mb-3">
+    <div class="w-14 h-14 rounded-2xl bg-primary-500 text-white flex items-center justify-center shadow-md mb-4">
+      <svg class="w-7 h-7" style="animation: ikspin 1.2s linear infinite" fill="none" stroke="currentColor" stroke-width="2.2" viewBox="0 0 24 24">
+        <path stroke-linecap="round" stroke-linejoin="round" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"/>
+      </svg>
+    </div>
+    <div class="font-serif text-xl md:text-2xl text-ink tracking-tight leading-tight">
+      {{ "Importing recipes…"|t }}
+    </div>
+    <p class="text-sm text-ink-2 mt-2 max-w-xs leading-relaxed">
+      {{ "Cross-checking ingredients and aisles"|t }}
+    </p>
+    <div class="text-[10px] font-mono uppercase tracking-widest text-ink-3 mt-4">
+      {{ "This may take a moment"|t }}
     </div>
   </div>
   <div ts-trigger="load delay 300ms" ts-req="/recipes/import/{{ id }}/status"></div>
