+++ templates/partials/recipes-share-all-button.html
@@ -1,0 +1,19 @@
+{% if has_shared %}
+<button ts-req="/recipes/make-all-private" ts-req-method="post"
+  class="inline-flex items-center gap-1.5 px-3 py-2 rounded-full border text-xs font-semibold cursor-pointer transition
+    bg-amber-50 border-amber-200 text-amber-700 hover:bg-amber-100">
+  <svg class="w-3.5 h-3.5" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+    <path stroke-linecap="round" stroke-linejoin="round" d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"/>
+  </svg>
+  {{ "Make All Private"|t }}
+</button>
+{% else %}
+<button ts-req="/recipes/share-all" ts-req-method="post"
+  class="inline-flex items-center gap-1.5 px-3 py-2 rounded-full border text-xs font-semibold cursor-pointer transition
+    bg-herb-50 border-herb-300 text-herb-700 hover:bg-herb-100">
+  <svg class="w-3.5 h-3.5" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+    <path stroke-linecap="round" stroke-linejoin="round" d="M3.055 11H5a2 2 0 012 2v1a2 2 0 002 2 2 2 0 012 2v2.945M8 3.935V5.5A2.5 2.5 0 0010.5 8h.5a2 2 0 012 2 2 2 0 104 0 2 2 0 012-2h1.064M15 20.488V18a2 2 0 012-2h3.064M21 12a9 9 0 11-18 0 9 9 0 0118 0z"/>
+  </svg>
+  {{ "Share All"|t }}
+</button>
+{% endif %}
