--- templates/partials/settings-recipes-share-button.html
@@ -1,11 +1,0 @@
-{% if has_shared %}
-<button ts-req="/settings/recipes/make-all-private" ts-req-method="post"
-  class="block w-full px-4 py-2 bg-amber-600 text-white font-semibold rounded-xl hover:bg-amber-700 text-center text-sm">
-  {{ "Make All Private"|t }}
-</button>
-{% else %}
-<button ts-req="/settings/recipes/share-all" ts-req-method="post"
-  class="block w-full px-4 py-2 bg-green-600 text-white font-semibold rounded-xl hover:bg-green-700 text-center text-sm">
-  {{ "Share All"|t }}
-</button>
-{% endif %}
