+++ templates/recipes-edit-instruction-row.html
@@ -1,10 +1,27 @@
-<div class="flex flex-col sm:flex-row gap-3 instruction-item">
-  <textarea name="instructions_description" rows="2" class="flex-1 min-w-0 px-2 sm:px-4 py-2 border border-line rounded-xl text-sm sm:text-base"
-    required></textarea>
-  <div class="flex gap-3">
-    <input type="number" name="instructions_time_next" min="0" value="0"
-      class="w-24 px-2 sm:px-4 py-2 border border-line rounded-xl text-sm sm:text-base" required />
-    <button ts-trigger="click" ts-action="remove 'parent .instruction-item'" type="button"
-      class="shrink-0 px-4 py-2 bg-red-100 text-red-600 rounded-xl hover:bg-red-200">✕</button>
+<div class="bg-paper border border-line-2 rounded-2xl p-3 md:p-4 flex gap-3 instruction-item">
+  <div class="w-9 h-9 rounded-xl flex items-center justify-center font-serif font-semibold bg-cream-2 text-ink shrink-0">
+    +
+  </div>
+  <div class="flex-1 min-w-0 space-y-2">
+    <textarea name="instructions_description" rows="2" placeholder="{{ "Describe this step..."|t }}" minlength="1"
+      class="w-full px-3 py-2 bg-cream border border-line rounded-lg text-sm text-ink leading-relaxed
+        focus:outline-none focus:border-primary-400 focus:ring-2 focus:ring-primary-100 transition resize-none" required></textarea>
+    <div class="flex items-center gap-2">
+      <svg class="w-3.5 h-3.5 text-ink-3 shrink-0" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+        <circle cx="12" cy="12" r="9"/><path stroke-linecap="round" d="M12 8v4l3 3"/>
+      </svg>
+      <input type="number" name="instructions_time_next" min="0" value="0"
+        class="w-16 px-2 py-1.5 bg-cream border border-line rounded-lg text-sm text-ink text-center font-mono
+          focus:outline-none focus:border-primary-400 focus:ring-2 focus:ring-primary-100 transition" required/>
+      <span class="text-xs font-mono uppercase tracking-wider text-ink-3">{{ "min"|t }}</span>
+      <div class="flex-1"></div>
+      <button ts-trigger="click" ts-action="remove 'parent .instruction-item'" type="button" title="{{ "Remove"|t }}"
+        class="shrink-0 w-8 h-8 flex items-center justify-center rounded-lg text-ink-3 hover:bg-red-50 hover:text-red-600 transition">
+        <svg class="w-4 h-4" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+          <path stroke-linecap="round" stroke-linejoin="round"
+            d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"/>
+        </svg>
+      </button>
+    </div>
   </div>
 </div>
