--- templates/settings-recipes.html
@@ -1,148 +1,0 @@
-{% extends "_settings.html" %}
-{% block title %}{{ "Recipes"|t }} - imkitchen{% endblock %}
-{% macro lazy_thumbnail(recipe, version, gradient) %}
-<div class="relative w-full h-full overflow-hidden">
-  <div class="absolute inset-0 bg-linear-to-r {{ gradient }} animate-pulse"></div>
-  <picture>
-    <source data-media="(max-width: 480px)"
-      data-srcset="/recipes/{{ recipe.id }}/thumbnail/mobile/image.webp?v={{ version }}" />
-    <img src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"
-      data-src="/recipes/{{ recipe.id }}/thumbnail/tablet/image.webp?v={{ version }}" alt="{{ recipe.name }}"
-      class="absolute inset-0 w-full h-full object-cover opacity-0 transition-opacity duration-500 [&.loaded]:opacity-100"
-      ts-trigger="closeby once" ts-action="lazy" />
-  </picture>
-</div>
-{% endmacro %}
-
-{% macro recipe_thumb(recipe, bg_gradient, placeholder_gradient, emoji) %}
-<div class="shrink-0 w-16 h-16 rounded-lg overflow-hidden bg-linear-to-br {{ bg_gradient }} flex items-center justify-center">
-  {% if let Some(version) = recipe.thumbnail_version %}
-  {{ lazy_thumbnail(recipe, version, placeholder_gradient) }}
-  {% else %}
-  <span class="text-2xl">{{ emoji }}</span>
-  {% endif %}
-</div>
-{% endmacro %}
-
-{% block sidebar_bottom %}
-
-{% if user.is_chef() %}
-<div class="hidden lg:block bg-paper rounded-xl shadow-md p-4">
-  <div class="space-y-2">
-    <div id="share-all-button">
-      {% if has_shared %}
-      <button ts-req="/settings/recipes/make-all-private" ts-req-method="post"
-        class="block w-full px-4 py-2 bg-amber-600 text-white font-semibold rounded-xl hover:bg-amber-700 text-center text-sm">
-        {{ "Make All Private"|t }}
-      </button>
-      {% else %}
-      <button ts-req="/settings/recipes/share-all" ts-req-method="post"
-        class="block w-full px-4 py-2 bg-green-600 text-white font-semibold rounded-xl hover:bg-green-700 text-center text-sm">
-        {{ "Share All"|t }}
-      </button>
-      {% endif %}
-    </div>
-  </div>
-</div>
-{% endif %}
-
-{% endblock %}
-
-{% block settings_content %}
-<div>
-  {% if user.is_chef() %}
-  <!-- Mobile Quick Actions -->
-  <div class="lg:hidden flex gap-2 mb-4">
-    <div id="share-all-button-mobile">
-      {% if has_shared %}
-      <button ts-req="/settings/recipes/make-all-private" ts-req-method="post"
-        class="flex-1 px-4 py-2 bg-amber-600 text-white font-semibold rounded-xl hover:bg-amber-700 text-center text-sm">
-        {{ "Make All Private"|t }}
-      </button>
-      {% else %}
-      <button ts-req="/settings/recipes/share-all" ts-req-method="post"
-        class="flex-1 px-4 py-2 bg-green-600 text-white font-semibold rounded-xl hover:bg-green-700 text-center text-sm">
-        {{ "Share All"|t }}
-      </button>
-      {% endif %}
-    </div>
-  </div>
-  {% endif %}
-
-  <!-- Filters -->
-  <div class="bg-paper rounded-xl shadow-md p-4 mb-6">
-    <form ts-trigger="change,input" ts-req="/settings/recipes" ts-req-selector="#recipes-list" ts-req-method="GET"
-      ts-req-strategy="last" ts-target="#recipes-list" class="grid md:grid-cols-3 gap-4" autocomplete="off">
-      <div>
-        <label class="block text-sm font-semibold text-ink-2 mb-2">{{ "Search"|t }}</label>
-        <input name="search" type="text" placeholder="{{ "Search by name, description or ingredient..."|t }}"
-          class="w-full px-4 py-2 border border-line rounded-xl focus:ring-2 focus:ring-primary-500" />
-      </div>
-      <div>
-        <label class="block text-sm font-semibold text-ink-2 mb-2">{{ "Recipe Type"|t }}</label>
-        <select name="recipe_type"
-          class="w-full px-4 py-2 border border-line rounded-xl focus:ring-2 focus:ring-primary-500">
-          <option value="">{{ "All Types"|t }}</option>
-          {% for variant in RecipeType::VARIANTS %}
-          <option value="{{ variant }}">{{ variant.as_ref()|t }}</option>
-          {% endfor %}
-        </select>
-      </div>
-      <div class="md:col-span-3 flex flex-wrap items-center gap-4">
-        <label class="flex items-center gap-2 cursor-pointer">
-          <input name="no_image" type="checkbox" value="true"
-            class="rounded border-line text-primary-500 focus:ring-primary-500" />
-          <span class="text-sm text-ink-2">{{ "No image"|t }}</span>
-        </label>
-      </div>
-    </form>
-  </div>
-
-  <!-- Recipe List -->
-  <div id="recipes-list" class="space-y-2">
-    {% if recipes.edges.is_empty() %}
-    <div class="bg-paper rounded-xl shadow-md p-12 flex flex-col items-center justify-center text-center">
-      <span class="text-6xl mb-4">🍽️</span>
-      <h3 class="text-xl font-semibold text-ink-2 mb-2">{{ "No recipes found"|t }}</h3>
-      <p class="text-ink-3">{{ "Try adjusting your filters or create a new recipe"|t }}</p>
-    </div>
-    {% else %}
-    {% for recipe in recipes.edges %}
-    <a id="recipe-{{ recipe.node.id }}" href="/recipes/{{ recipe.node.id }}" {% if let (true, true,
-      Some(cursor))=(loop.last, recipes.page_info.has_next_page, recipes.page_info.end_cursor.to_owned()) %}
-      ts-req="/settings/recipes?after={{ cursor.to_string() }}{% if let Some(recipe_type) = query.recipe_type %}&recipe_type={{ recipe_type }}{% endif %}{% if let Some(true) = query.no_image %}&no_image=true{% endif %}"
-      ts-req-method="GET" ts-req-selector="children #recipes-list" ts-swap="afterend" ts-trigger="visible once" {% endif
-      %} class="group flex items-center gap-4 bg-paper rounded-xl shadow-sm hover:shadow-md p-3 transition-all">
-      {% match recipe.node.recipe_type.0 %}
-      {% when RecipeType::MainCourse %}
-      {% call recipe_thumb(recipe.node, "from-orange-400 to-red-500", "from-orange-400 via-orange-200 to-red-400", "🍛") %}{% endcall %}
-      {% when RecipeType::Appetizer %}
-      {% call recipe_thumb(recipe.node, "from-blue-400 to-blue-500", "from-blue-400 via-blue-200 to-blue-400", "🥗") %}{% endcall %}
-      {% when RecipeType::Dessert %}
-      {% call recipe_thumb(recipe.node, "from-pink-400 to-pink-500", "from-pink-400 via-pink-200 to-pink-400", "🍰") %}{% endcall %}
-      {% when RecipeType::Accompaniment %}
-      {% call recipe_thumb(recipe.node, "from-purple-400 to-purple-500", "from-purple-400 via-purple-200 to-purple-400", "🍚") %}{% endcall %}
-      {% when RecipeType::Beverage %}
-      {% call recipe_thumb(recipe.node, "from-cyan-400 to-cyan-500", "from-cyan-400 via-cyan-200 to-cyan-400", "🥤") %}{% endcall %}
-      {% when RecipeType::Condiment %}
-      {% call recipe_thumb(recipe.node, "from-amber-400 to-amber-500", "from-amber-400 via-amber-200 to-amber-400", "🥫") %}{% endcall %}
-      {% endmatch %}
-      <div class="min-w-0 flex-1">
-        <h3 class="font-semibold text-ink truncate group-hover:text-primary-600 transition-colors">{{ recipe.node.name }}</h3>
-        <div class="flex items-center gap-3 text-xs text-ink-3 mt-1">
-          <span>{{ recipe.node.recipe_type.as_ref()|t }}</span>
-          <span>⏱️ {{ (recipe.node.cook_time + recipe.node.prep_time)|minutes }}</span>
-          {% if recipe.node.is_shared %}
-          <span class="text-green-600 font-medium">{{ "Shared"|t }}</span>
-          {% endif %}
-        </div>
-      </div>
-      <svg class="shrink-0 w-5 h-5 text-ink-3 group-hover:text-primary-500 transition-colors" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-        <path stroke-linecap="round" stroke-linejoin="round" d="M9 5l7 7-7 7" />
-      </svg>
-    </a>
-    {% endfor %}
-    {% endif %}
-  </div>
-</div>
-{% endblock %}
