+++ web/recipe/src/routes/index.rs
@@ -1,4 +1,7 @@
-use axum::{extract::State, response::IntoResponse};
+use axum::{
+    extract::State,
+    response::{IntoResponse, Redirect},
+};
 use axum_extra::extract::Query;
 use evento::cursor::{Args, ReadResult, Value};
 use imkitchen_core::recipe::query::user::{RecipesQuery, SortBy, UserViewList};
@@ -9,9 +12,11 @@ use strum::VariantArray;
 
 use imkitchen_web_shared::{
     AppState,
-    auth::AuthUser,
+    auth::{AuthUser, RequireChef, RequirePremium},
     template::{Template, filters},
 };
+
+use super::detail::SetUsernameModalTemplate;
 
 #[derive(askama::Template)]
 #[template(path = "recipes-index.html")]
@@ -20,6 +25,7 @@ pub struct IndexTemplate {
     pub user: AuthUser,
     pub recipes: ReadResult<UserViewList>,
     pub query: PageQuery,
+    pub has_shared: bool,
 }
 
 impl Default for IndexTemplate {
@@ -29,6 +35,7 @@ impl Default for IndexTemplate {
             user: AuthUser::default(),
             recipes: ReadResult::default(),
             query: Default::default(),
+            has_shared: false,
         }
     }
 }
@@ -43,6 +50,9 @@ pub struct PageQuery {
     pub search: Option<String>,
     pub sort_by: Option<SortBy>,
     pub in_meal_plan: Option<bool>,
+    pub mine: Option<bool>,
+    pub no_image: Option<bool>,
+    pub view: Option<String>,
 }
 
 #[tracing::instrument(skip_all, fields(user = user.id))]
@@ -66,8 +76,9 @@ pub async fn page(
         .and_then(|v| RecipeType::from_str(v.as_str()).ok());
 
     let in_meal_plan = input.in_meal_plan.unwrap_or(false);
+    let mine = input.mine.unwrap_or(false);
 
-    let dietary_restrictions = if in_meal_plan {
+    let dietary_restrictions = if in_meal_plan || mine {
         vec![]
     } else {
         let preferences = imkitchen_web_shared::try_page_response!(
@@ -77,16 +88,40 @@ pub async fn page(
         preferences.dietary_restrictions
     };
 
+    let (user_id, is_shared) = if mine {
+        (Some(user.id.to_owned()), None)
+    } else {
+        (None, Some(true))
+    };
+
+    // `in_meal_plan: Some((_, false))` is an exclusion (NOT EXISTS), used by the
+    // community browse default to hide already-planned recipes from the picker.
+    // When filtering by ownership, that exclusion would hide every recipe the
+    // user has already added — drop the plan filter unless In plan is on.
+    let in_meal_plan_filter = if in_meal_plan {
+        Some((user.id.to_owned(), true))
+    } else if mine {
+        None
+    } else {
+        Some((user.id.to_owned(), false))
+    };
+
+    let has_thumbnail = if input.no_image.unwrap_or(false) {
+        Some(false)
+    } else {
+        None
+    };
+
     let recipes = imkitchen_web_shared::try_page_response!(
         app.core.recipe.filter_user(RecipesQuery {
             exclude_ids: None,
-            user_id: None,
+            user_id,
             recipe_type,
-            is_shared: Some(true),
-            has_thumbnail: None,
+            is_shared,
+            has_thumbnail,
             dietary_restrictions,
             dietary_where_any: false,
-            in_meal_plan: Some((user.id.to_owned(), in_meal_plan)),
+            in_meal_plan: in_meal_plan_filter,
             sort_by: input.sort_by.unwrap_or_default(),
             args: args.limit(20),
             search: input.search,
@@ -94,12 +129,78 @@ pub async fn page(
         template
     );
 
+    // Drives the Share All / Make All Private toggle on the chef toolbar.
+    // Only considers the loaded page of recipes.
+    let has_shared = recipes.edges.iter().any(|r| r.node.is_shared);
+
     template
         .render(IndexTemplate {
             user,
             recipes,
             query,
+            has_shared,
             ..Default::default()
         })
+        .into_response()
+}
+
+#[tracing::instrument(skip_all, fields(user = user.id))]
+pub async fn create(
+    template: Template,
+    RequirePremium(user): RequirePremium,
+    State(app): State<AppState>,
+) -> impl IntoResponse {
+    let id = match imkitchen_web_shared::try_response!(anyhow: app.core.recipe.find_user_draft(&user.id), template)
+    {
+        Some(id) => id,
+        _ => imkitchen_web_shared::try_response!(
+            app.core.recipe.create(&user.id, user.username.to_owned()),
+            template
+        ),
+    };
+
+    Redirect::to(&format!("/recipes/{id}/edit")).into_response()
+}
+
+#[derive(askama::Template)]
+#[template(path = "partials/recipes-share-all-button.html")]
+pub struct ShareAllButtonTemplate {
+    pub has_shared: bool,
+}
+
+#[tracing::instrument(skip_all, fields(user = user.id))]
+pub async fn share_all(
+    template: Template,
+    State(app): State<AppState>,
+    RequireChef(user): RequireChef,
+) -> impl IntoResponse {
+    let Some(ref username) = user.username else {
+        return (
+            [("ts-swap", "skip")],
+            template.render(SetUsernameModalTemplate),
+        )
+            .into_response();
+    };
+
+    imkitchen_web_shared::try_response!(
+        app.core.recipe.share_all_to_community(&user.id, username),
+        template
+    );
+
+    template
+        .render(ShareAllButtonTemplate { has_shared: true })
+        .into_response()
+}
+
+#[tracing::instrument(skip_all, fields(user = user.id))]
+pub async fn make_all_private(
+    template: Template,
+    State(app): State<AppState>,
+    RequireChef(user): RequireChef,
+) -> impl IntoResponse {
+    imkitchen_web_shared::try_response!(app.core.recipe.make_all_private(&user.id), template);
+
+    template
+        .render(ShareAllButtonTemplate { has_shared: false })
         .into_response()
 }
