+++ web/settings/src/routes/mod.rs
@@ -2,4 +2,3 @@ pub mod account;
 pub mod billing;
 pub mod general;
 pub mod invoices;
-pub mod recipes;
